Still updating

+++ src/models/phase.py
@@ -14,7 +14,7 @@ class Message(BaseModel):
 
 class Execution(BaseModel):
     commands: List[str] = Field(default_factory=list)
-    durations: Optional[List[int]] = Field(default_factory=list)
+    durations: Optional[List[float]] = Field(default_factory=list)
     traceback: Optional[str] = Field(default=None)
     feedback: Optional[Message] = Field(default=None)
     response: Optional[Message] = Field(default=None)
@@ -22,7 +22,7 @@ class Execution(BaseModel):
 
 class PromptPhase(BaseModel):
     responses: Optional[List[str]] = Field(default_factory=list)
-    durations: Optional[List[int]] = Field(default_factory=list)
+    durations: Optional[List[float]] = Field(default_factory=list)
     tracebacks: Optional[List[str]] = Field(default_factory=list)
     execution: Optional[Execution] = Field(default=None)
